Minor changes to work with new C client library

+++ src/server/gui/GridLayout.h
@@ -13,11 +13,11 @@ class Container;
 class GridLayout : public LayoutManager {
 public:
     GridLayout(int32_t cols=1, int32_t rows=1);
-    GridLayout(std::vector<int16_t> x_props, std::vector<int16_t> y_props);
+    GridLayout(std::vector<uint8_t> x_props, std::vector<uint8_t> y_props);
     void place(Container& parent, Widget& child);
     void reset();
 private:
-    std::vector<int16_t> x_props, y_props;
+    std::vector<uint8_t> x_props, y_props;
     int32_t next_col = 0, next_row = 0, next_x = 0, next_y = 0;
 };
